agent/requirements.txt
fastapi[standard]
groq
python-dotenv
pydantic